open_mahjong_server/server/requirements.txt
fastapi
uvicorn
websockets
psycopg2-binary
PyJWT
cryptography 